photon_pt_spectrum: fix bug in taking the ratio

+++ MainAnalysis/src/compare_photon_pt_spectrum_accumulate.C
@@ -91,7 +91,7 @@ int compare_photon_pt_spectrum(char const* config, char const* selections, const
     normalise_to_unity(photon_pt_data, photon_pt_mc);
 
     auto photon_pt_ratio = new history<TH1F>("photon_pt_ratio"s, "", fpt, ihf->size());
-    for (size_t i = 0; i < ihf->size(); ++i) {
+    for (int64_t i = 0; i < ihf->size(); ++i) {
         (*rho_ratio)[i]->Divide((*photon_pt_data)[i], (*photon_pt_mc)[i]);
     }
